+++ server/src/features/ingredient/routes.js
@@ -1,29 +1,29 @@
 const express = require('express');
-const helper = require('./helper');
+const service = require('./service');
 
 const router = express.Router();
 
-router.get('/', (req, res) => helper.findAll()
+router.get('/', (req, res) => service.findAll()
   .then((data) => res.status(200).send(data))
   .catch((err) => res.status(500).send(err)));
 
-router.get('/:id', (req, res) => helper.findById(req.params.id)
+router.get('/:id', (req, res) => service.findById(req.params.id)
   .then((data) => res.status(200).send(data))
   .catch((err) => res.status(500).send(err)));
 
-router.post('/', (req, res) => helper.create(req.body)
+router.post('/', (req, res) => service.create(req.body)
   .then((data) => res.sendStatus(200).send(data))
   .catch((err) => res.sendStatus(500).send(err)));
 
-router.put('/:id', (req, res) => helper.update(req.params.id, req.body)
+router.put('/:id', (req, res) => service.update(req.params.id, req.body)
   .then((data) => res.status(200).send(data))
   .catch((err) => res.status(500).send(err)));
 
-router.delete('/', (req, res) => helper.destroyAll()
+router.delete('/', (req, res) => service.destroyAll()
   .then((data) => res.status(200).send(data))
   .catch((err) => res.status(500).send(err)));
 
-router.delete('/:id', (req, res) => helper.destroy(req.params.id)
+router.delete('/:id', (req, res) => service.destroy(req.params.id)
   .then((data) => res.status(200).send(data))
   .catch((err) => res.status(500).send(err)));
